requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.22.0
bcrypt==3.1.3
billiard==3.3.0.23
celery==3.1.25
Cerberus==0.9.2
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cryptography==2.0.3
enum34==1.1.6
Eve==0.7.4
Events==0.2.2
Fabric==1.13.1
Flask==0.12.2
Flask-PyMongo==0.5.1
idna==2.5
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
kombu==3.0.37
MarkupSafe==0.23
paramiko==2.2.1
pyasn1==0.3.2
pycparser==2.18
pymongo==3.5.0
PyNaCl==1.1.2
python-ldap==2.4.41
pytz==2017.2
requests==2.18.3
simplejson==3.11.1
six==1.10.0
urllib3==1.22
Werkzeug==0.11.15
